Solution to Find Maximum of Three Numbers

+++ Find Maximum of Three Numbers@@6788dd63490637cfc44e1b94/code.c
@@ -1,17 +1,17 @@
 #include <stdio.h>
 
 int main() {
-    float num1, num2, num3;
+    int num1, num2, num3;
 
-    scanf("%f %f %f", &num1, &num2, &num3);
+    scanf("%d %d %d", &num1, &num2, &num3);
 
     // Find the maximum number
     if (num1 >= num2 && num1 >= num3) {
-        printf("%.2f\n", num1);
+        printf("%d\n", num1);
     } else if (num2 >= num1 && num2 >= num3) {
-        printf("%.2f\n", num2);
+        printf("%d\n", num2);
     } else {
-        printf("%.2f\n", num3);
+        printf("%d\n", num3);
     }
 
     return 0;
